Add interactive kernel shell with VGA + keyboard support
- Create console driver (kernel/drivers/console.c) for dual serial+VGA output
- Create kernel shell (kernel/drivers/kshell.c) with built-in commands:
  help, echo, clear, ps, uptime, mem, version, reboot
- Wire irq_init() into kernel_main (was never called - no HW interrupts worked)
- Initialize VGA console and PS/2 keyboard during boot
- Migrate boot messages from serial-only to console (serial + VGA)
- Replace hardcoded test tasks with interactive kshell_run() loop
- NexusOS now boots to a live shell prompt in QEMU

+++ kernel/drivers/ata.h
@@ -15,8 +15,8 @@ typedef struct {
 } ata_drive_info_t;
 
 void ata_init(void);
-int ata_read_sectors(uint32_t lba, uint16_t count, void *buffer);
-int ata_write_sectors(uint32_t lba, uint16_t count, void *buffer);
+int ata_read_sectors(uint32_t lba, uint32_t count, void *buffer);
+int ata_write_sectors(uint32_t lba, uint32_t count, void *buffer);
 ata_drive_info_t *ata_get_drive_info(void);
 
 #endif
